requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
fastapi==0.110.0
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.4
packaging==24.0
passlib==1.7.4
platformdirs==4.2.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
sentry-sdk==2.11.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.2.2
uvicorn==0.28.0
virtualenv==20.25.0
websockets==12.0
